+++ packages/splat-transform/src/tasks/SplitSplatTask.ts
@@ -1,4 +1,4 @@
-import { splitSplat } from '../native/index.js';
+import { SplitNormal, splitSplat } from '../native/index.js';
 import type { SplatData } from '../SplatData.js';
 import { type Context, BaseTask, type SingleFile } from './BaseTask.js';
 
@@ -6,18 +6,23 @@ export interface Config {
     input: string;
     output: string;
     type: string;
+    splitNormal?: SplitNormal | keyof typeof SplitNormal;
     blockPrecision?: number;
 }
 
 export class SplitSplatTask extends BaseTask<Config> {
     override exec(config: Config, { logger, resources }: Context) {
-        const { input, output, type, blockPrecision = 0.5 } = config;
+        const { input, output, type, blockPrecision = 0.5, splitNormal = SplitNormal.None } = config;
         const splat = resources.get(input) as SplatData;
         logger.info(`loaded -> "${input}"`);
         logger.info(`block precision -> ${blockPrecision}`);
         logger.info('splitting splat');
         logger.time('split elapsed');
-        const { splats } = splitSplat(splat, blockPrecision);
+        const { splats } = splitSplat(
+            splat,
+            blockPrecision,
+            typeof splitNormal === 'string' ? SplitNormal[splitNormal] : splitNormal,
+        );
         logger.timeEnd('split elapsed');
         const outputs: SingleFile[] = [];
         for (let i = 0; i < splats.length; i++) {